% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_df}
\alias{map_df}
\title{map data combined with NYC tract id in simple features}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2166 rows and 7 columns.
}
\usage{
map_df
}
\description{
map data combined with NYC tract id in simple features
}
\keyword{datasets}
