% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{PlotManager}
\alias{PlotManager}
\title{R6 class for plotting}
\description{
R6 class for plotting

R6 class for plotting
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-plot_dist}{\code{PlotManager$plot_dist()}}
\item \href{#method-clone}{\code{PlotManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_dist"></a>}}
\if{latex}{\out{\hypertarget{method-plot_dist}{}}}
\subsection{Method \code{plot_dist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotManager$plot_dist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
